(maint) Move bash scripts to resources and pull them in directly
Previously, we were curling to github to pull down bash scripts from this repo. This is no longer viable, since the repo is no longer public, and is not necessary since the runner already clones this repo. This modifies the groovy scripts to pull in the bash script directly using libraryResource, which also requires moving the bash scripts into the 'resources' directory.